requirements.txt
polars
pyarrow
fastapi
uvicorn
sentence-transformers
faiss-cpu
httpx
pydantic
python-dotenv
tqdm
jinja2